Remove trailing spaces

+++ Activity_8/src/temperatureconverter/FahrenheitToCelsius.java
@@ -6,16 +6,13 @@ public class FahrenheitToCelsius {
 
 	public static void main(String[] args) {
 		Scanner input = new Scanner(System.in);
-		
+
 		System.out.print("Enter temperature in Fahrenheit: ");
-		double fahrenheit = input.nextDouble();
-		
-		double celcius = ( fahrenheit - 32 ) * 5 / 9;
-		
+		float fahrenheit = input.nextFloat();
+
+		float celcius = ( fahrenheit - 32 ) * 5 / 9;
+
 		System.out.printf("Tempereature in Celcius: %.2f°C%n", celcius);
-		
 		input.close();
 	}
-
 }
-
